requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.4
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.5
command-not-found==0.3
contourpy==1.0.7
cryptography==40.0.1
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
distro==1.7.0
distro-info==1.1+ubuntu1
duplicity==0.8.22
fasteners==0.17.3
fonttools==4.39.3
future==0.18.2
html5lib==1.1
httplib2==0.20.4
idna==3.4
importlib-metadata==4.6.4
jeepney==0.8.0
joblib==1.2.0
keyring==23.8.2
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.22.0
lxml==4.9.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.7.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
packaging==23.0
pandas==1.5.3
paramiko==2.10.4
patsy==0.5.3
pexpect==4.8.0
Pillow==9.4.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGithub==1.58.1
PyGObject==3.42.2
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-apt==2.4.0+22.10
python-dateutil==2.8.2
python-debian==0.1.46
python-dotenv==1.0.0
python-magic==0.4.26
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.11
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==3.3.3
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.2
statsmodels==0.13.5
systemd-python==235
threadpoolctl==3.1.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.15
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
webencodings==0.5.1
wrapt==1.15.0
xdg==5
xkit==0.0.0
zipp==1.0.0
